requirements.txt
openai>=1.0.0
google-maps-services>=4.10.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.10.0
torch>=2.0.0
open-clip-torch>=2.20.0
pandas>=2.0.0
folium>=0.16.0
tqdm>=4.66.0
python-dotenv>=1.0.0
easyocr>=1.7.0
shapely>=2.0.0
